requirements.txt
tornado~=6.0.4
requests==2.24.0
selenium==3.141.0
python_dotenv==0.14.0
